types: improve code formatting and type consistency
Refactor type definitions and function signatures across the
OutbreakDetectionCore module to improve readability and maintain
consistency with Runic formatting standards. Changes include:

- Break long export statements across multiple lines for clarity
- Split multi-parameter function signatures with one parameter per
  line
- Reformat LightSumTypes.@sumtype macro calls to single-line format
- Add explicit AutoHashEquals import to main module
- Remove unused function parameters (e.g., `noise` parameter in
  get_noise_description)
- Simplify type annotations (e.g., `TMethod` to
  `OptimizationMethods`)
- Remove unnecessary `where` clause from
  calculate_combination_to_run
- Improve keyword argument formatting in function calls

Files modified:
- OutbreakDetectionCore.jl: Add AutoHashEquals import
- scenario-creation.jl: Reformat function signatures and type
  annotations
- dynamics-parameters.jl: Split export and function signatures
- noise-specifications.jl: Reformat exports and remove unused params
- optimization-results.jl: Remove redundant AutoHashEquals import
- optimization-types.jl: Split export statements
- outbreak-specifications.jl: Format exports and function signatures
- scenario-parameters.jl: Improve formatting consistency
- scenario-specifications.jl: Split long exports
- simulation-results.jl: Format function signatures
- state-parameters.jl: Improve export formatting
- test-specifications.jl: Split long exports
- time-parameters.jl: Format exports

This is a structural change with no behavioral modifications. All
changes maintain backward compatibility and improve code
maintainability.

+++ OutbreakDetectionCore/src/types/test-specifications.jl
@@ -1,5 +1,8 @@
-export IndividualTestSpecification, TestPositivity,
-    get_test_description, table_test_type, plot_test_description
+export IndividualTestSpecification,
+    TestPositivity,
+    get_test_description,
+    table_test_type,
+    plot_test_description
 
 """
     IndividualTestSpecification
@@ -47,7 +50,11 @@ end
 
 Get a table-formatted test type description.
 """
-function table_test_type(sensitivity, specificity, test_lag)
+function table_test_type(
+        sensitivity::Float64,
+        specificity::Float64,
+        test_lag::Int64
+    )
     return if sensitivity == 1.0 && specificity == 0.0 && test_lag == 0
         "Clinical Case Definition"
     elseif sensitivity == specificity && sensitivity < 1.0 && test_lag == 0
